requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
lxml==5.3.2
multidict==6.4.3
propcache==0.3.1
PyMuPDF==1.25.5
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
